Set options.request.agent as actions.js default instead of plugin
This removes the code that hooks into the request lifecycle and instead
just sets the agent as a default based on the HTTPS_PROXY environment
variable when the Octokit class is instantiated.

Setting the options.request.agent field in the plugin had an issue where
the agent field wasn't being passed to the fetch wrapper in core.js,
since the request options were being set before the plugins were run.

The reason the defaults are passed as a function value instead of an
object is that it was difficult to test different behavior based on env
variables in src/index.ts because it was only run once during the import
at the top of the test file. Using the function lets us change the
environment variable in between test cases and lets the change get
reflected when octokit is instantiated, and not when it's imported.